refactor: divide useAuth into separate hooks
